Move the initialization of the Maps SDK to BaseActivity and set the preferred renderer to LATEST
The initialization of the Maps SDK happens only once, so in order to force it to use the new Renderer, the initialization needs to happen before any other Map elements are created. This change also removes the callback interface OnMapsSdkInitializedCallback as the API already logs details about the Renderer being used, making this unnecessary.